create evetns module, remove unused connection manager events and related tests
